FULL EOG module but need improvement.
This is the original altium file module Amplifier for EMG, it needs some modification to make EOG, actually the amplification process in EOG and EMG. it's the same process, but there's some part of devices that be removed. so if you want to make this EOG device use this Altium file and then remove another part // just the same as EOG module schematic picture that you can see in another file.